♻️ refactor: 홈 화면 컴포넌트 이동 - /components/home -> /features/home/components - deadcode 삭제
