--- src/components/home/MemeListContainer/DropDown/index.ts
@@ -1,1 +1,0 @@
-export * from "./MemeSortDropDown";
