--- src/components/home/MemeListContainer/MemeList/index.ts
@@ -1,2 +1,0 @@
-export * from "./CommonMemeList";
-export * from "./UserFindMemeList";
