+++ src/components/home/Skeleton/index.ts
@@ -1,2 +1,1 @@
-export * from "./SkeletonMeme";
 export * from "./SkeletonTagList";
