--- src/components/home/UserSharedMeme/index.ts
@@ -1,1 +1,0 @@
-export * from "./UserSharedMemeList";
