+++ src/features/home/components/MemeListContainer/index.ts
@@ -1,2 +1,1 @@
 export * from "./MemeListContainer";
-export * from "./UserSharedMeme";
