+++ src/pages/index.tsx
@@ -5,7 +5,7 @@ import type { NextSeoProps } from "@/common/components/NextSeo";
 import { NextSeo } from "@/common/components/NextSeo";
 import { PullToRefresh } from "@/common/components/PullToRefresh";
 import { DEFAULT_DESCRIPTION, SITE_NAME } from "@/common/utils";
-import { MemeListContainer } from "@/components/home";
+import { MemeListContainer } from "@/features/home/components";
 
 const HomePage: NextPage = () => {
   return (
